Penamabahan status ganti password pada get profile untuk team, mempastikan peresntase di milestone berupa int, fix update profile team tidak bisa valid

+++ app/Http/Services/Controller/InsertBuatMilestoneService.php
@@ -38,7 +38,7 @@ class InsertBuatMilestoneService
             'topik' => $request->topik,
             'deskripsi' => $request->deskripsi,
             'tautan' => $request->tautan,
-            'persentase' => abs($request->persentase),
+            'persentase' => (int)abs($request->persentase),
             'tanggal_tegat' => $request->tanggal_tegat,
             'status' => 0,
             'waktu_buat' => new \DateTime(),
